ai_service/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
requests
pillow
groq
httpx